+++ backend/src/main/java/com/illch/auth/infrastructure/dto/OauthTokenRequest.java
@@ -1,4 +1,4 @@
-package com.illch.infrasturcture.auth;
+package com.illch.auth.infrastructure.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
