--- backend/src/main/java/com/illch/infrasturcture/auth/JwtToken.java
@@ -1,10 +1,0 @@
-package com.illch.infrasturcture.auth;
-
-import lombok.Getter;
-
-@Getter
-public abstract class JwtToken {
-
-    private String secretKey;
-    private Long validTime;
-}
